requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic>=2.8.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
mangum==0.17.0
sqlalchemy==2.0.25
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0
